added paranoia inputs on I2C because bitgen config options get lost
Although the "right" solution by xilinx is to configure bitgen
to have -g UnusedPin:PullNone, the method to configure this isn't
very portable especially through github. So in ample paranoia, we've
simply made the unused I2C pins inputs and tied them to the FPGA
LED flashing pin, which means the LED will flicker a bit if
I2C is being used.